Initial condition cannot depend on boundcond.f90, because boundcond already depends on special and chemistry (why?), which would caused a circular dependency.
I think that boundcond should be a low level module, while right now 
it depends even on Viscosity. Ideally, one should be capable 
of updating the boundaries to code an initial condition (one that,
for instance, requires calculation of gradients accross the boundary).   

Commented out the call to update_ghosts in the sample initial 
condition file.