Commented out about 15 arrays of size (mx,my,mz). This will probably break some of the chemistry tests. But we are supposed to run a MEMORY OPTIMISED code here.
These subs will have to be pencilised, or moved to someone's special file.

Indentation, comments, and variable declarations are not even close to being
readable to anyone.

People, this is a code shared by all of us. Could you *please* follow the
standards of the rest of the Pencil Code? Otherwise the chemistry stuff will
have to split off in its own code.