Added a temporary pencil in cdata.f90 to fix a pointer pb with gfortran (and also g95):
I want to add an implicit sample that runs under both gfortran (default
compiler) and Intel compiler ifort. But there is a pb with gfortran when
one try to share a pencil hcond(mx) between temperature_idealgas and
boundcond modules for the 'c3' boundary condition (imposed flux with
variable hcond at the bottom). This pointer pb leads to random values
for the pencil hcond() when one gets it using get_shared_variable()
in boundcond.  Nevertheless this pb does not exist with ifort.

Therefore the trick consisting in adding the hcond(mx) pencil in cdata
in *really* temporary.