**** in Makefile.src:
The fact that gfortran is now the default compiler leads to an option
pb when one runs a double precision sample as the default flag is:

FFLAGS_DOUBLE=-fdefault-real-8

and this option is of course not recognized by ifort or g95. If one uses
gfortran, no pb, but if one uses the settings for a specific machine
(i.e. the settings _below_ this default flag), the pb arises.

The solution is to add the following line at the beginning of its own
specific settings (for example for my PC shal running ifort):

#include $(PENCIL_HOME)/src/Makefile_ifort.incl #(shal)

In this include file, the FFLAGS_DOUBLE flag is set again to

FFLAGS_DOUBLE=-r8

and thus the default gfortran flag is overwritten. Wolfgang wrote 3
include files like that for the three common compilers:

Makefile_g95.incl
Makefile_gfortran.incl
Makefile_ifort.incl

Some people already do that in their specific settings but the majority
do not.

*** in temperature_idealgas: removed the optional dhcond argument in the call to heatcond_TT() when not needed

*** in README: added the 1d-tests/implicit_diffusion sample in the list