sed -i 's/frak/frac/g'
Because the code should be in English :-) 

BTW, this subroutine (that sets the shear-periodic boundary condition), 
deserves some commenting, references, etc. It's sorta cryptical the way it 
is done. It's not even clear who coded them since the headers say "nils" in 
initiate_shearing, "ulf" (Torkelsson?) in finalize_shearing, but it's Anders 
and Axel who are all over the annotate and log.