Changed
  gamma1  -> gamma_m1  (gamma-1)
  gamma11 -> gamma_inv (1/gamma)

The openly stated goal is to have gamma_inv replaced by gamma1, since 
everywhere else in the code quantity1 means 1/quantity. But as for 
now people are used to have gamma1=gamma-1, some time is needed for 
this habit to fall out of use. So we shall live with gamma_inv for 
some time, and in perhaps 6 months we can change it to gamma1. 

Decision taken in the Pencil Code Meeting 2009 as suggested by me 
(change gamma1 and gamma11) and Wolfgang (have a transitional gamma_inv 
to reduce the trauma).