Fixed usage of optional argument ladd, which was incorrectly coded in many subroutines. Remember: when you assign a value in the declaration of a variable like this:
  logical :: ladd1 = .false.

then the value is assigned only the first time the subroutine is called
(i.e. this is equvalent to
  logical, save :: ladd1 = .false.
).
This is true for at least some compilers, but it is probably even
required by the F90 standard.

I am elaborating on that because most of us (certainly myself) got this
wrong occasionally.