Created by MR. This directory is devoted to bugs as a consequence of the 'undesired save attribute' of local (automatic) variables: An initialisation of such variables inside their declaration causes the compiler to give them silently the attribute "save", hence they become static and preserve their values across procedure calls. The initialization is performed only once. This is in most cases not what the programmer intended and sometimes dangerous.
The file "list" is the result of a scan of $PENCIL_HOME/src for such situations, ordered alphabetically by the name of the .f90
source. The separating line is "<filename>: contains".
Note that the scan is potentially incomplete as it doesn't cover continuation lines.

Sometimes, by its semantics, the affected variable will turn out to be actually a symbolic constant or to deserve to be stored
statically. In these cases I recommend to set the corresponding attributes "parameter" and "save" explicitly.
In all other cases the initialization should be replaced by an assignment or, if possible, simply deleted.