Added a "gaussian-pos" option for initial chemical concentrations, corresponding to a punctal initial source of compounds.
Removed at the same time a strange behavior for "lone_spec" condition,
i.e. when there is only one chemical compound, forcing a uniform 
concentration in the system... It is actually causing a problem when
working with chemical systems with one unique compound, as it becomes 
impossible to have a non-homogeneous system. What could be the reason
for such a behavior ???