The slices have been implemented. This should be now functional, so that you can start to play and experiment it (mostly the implementation of the var files have to be done)... so please try it :).
The most simple is to use the hdf5.py module to create the hdf5 file one
your runs, and then to try to access your data through this file. This is
very straightforward, at least in python, where you just have to open the
HDF5 file in order to have access to the data as if they were regular numpy
array.