Block domain decomposition of particles has landed:
In block domain decomposition the main mesh is divided among the processors in
the usual way. The local mesh is then divided into so-called "bricks", small
volumes of grid points. Particles are counted in each of those bricks, and then
the bricks are distributed among the processors so that each processor has
approximately the same number of particles. A brick fostered by a processor is
called a "block".

In each time-step the relevant dynamical variables must be transferred from
bricks at the parent processors to blocks at the foster processors. This can be
e.g. gas velocity field (for drag force) or gravitational potential (for
self-gravity).

A processor can open up new bricks in its own domain, if a particle moves into
an empty brick. Full load balancing is performed at regular intervals.  Here
each processor count particles in their blocks and sends the information to the
parent processors. The parent processors then decide on a new division of
particle blocks.

Preliminary results show that this does slow the code down (by a factor 2-3),
but since the code can slow down by factors more than 10 if particles cluster
on a single CPU, there is still a lot of potential for speed up.

I also need to look at optimizing block domain decomposition for speed.