Allocate all 1-D and 2-D arrays used for diagnostics. This follows the example made by Dhruba for x-averages and z-averages.
I think this is much more fair, e.g. that someone who does not use cylindrical
averages should not have to consider an array of size
(nrcyl,0:nz,nprocz,mnamerz). This array is what gave me trouble compiling a
1024^3 simulation. It should be possible to abolish the nprocz dimension with
some work, but I know too little about cylindrical averages to do that. That
also means that I could not check whether they still work. Could someone do
that?

Ideally the code should return an error if the arrays can not be allocated, but
that may depend on the compiler.