Fixed the auto-tests 2d-tests/globaldisc and mdwarf that crashed since some days. It was due to the fact that the phi_averages variables can have shorthand labels declared in the file 'phiaver.in' (e.g. uumphi means urmphi, upmphi and uzmphi) and that now these fields are allocated dynamically.
As a consequence, the number of lines in 'phiaver.in' was no longer the
pertinent value to calculate the total size of the cnamerz() and cformrz()
arrays... Now a local variable in register.f90 takes into account the
occurence of, e.g., uumphi and increases accordingly nnamerz _before_
the array allocation, and then come back to the former value before the
call to expand_name().

Boris.