Implemented two-way drag forces in particle blocks. This of course is only needed for multi-processor simulations.
Since the infrastructure is quite different from the usual pencil loop, I made
a whole new version of particles_dust, particles_dust_blocks.f90.

It gives the same results as with the usual domain decomposition, but with a
gaussian distribution of particles the new method is already significantly
faster at four CPUs.