Fixed path to pc_update_autotest.sh, but not sure whether this will work for the cron job (I got   Can't exec "pc_update_autotest.sh": No such file or directory at   ./bin/auto-test line 390. and don't have the time to test whether my changes help.
[I will be away until 1 January, with possibly just access to email
(through the webmail interface), so I might forward error messages if
things still don't work, but am unlikely to fix anything myself.]

In my opinion, the fact that a script from bin/ calls a script from utils/
points to a dubious design. But until I come up with a better one, it
should stay in place.