Trying to fix pencil-check related problem in chemistry. On 2 occasions 2 captured problems, and those occasions are now outputted. For example, I get:
  get_reaction_rate: iproc,Kc= 0 1.5913898115134767E-30 0. 0. 0. 0. 0. 

etc, but Kc must not be negative, because we divide by it. Likewise,

  get_reaction_rate: iproc,dSR_dHRT= 0 65.38927071432762 23482.647058488285

etc, we get overflows when exp(dSR_dHRT) > 1e300 or so, and this is now
way over. After have captured these occurences, I ended up with an overflow
in the kf=... line below

!
!  Find forward rate constant for reaction 'reac'
!
        kf(:)=B_n(reac)*p%TT(:)**alpha_n(reac)*exp(-E_an(reac)/Rcal/p%TT(:))
        if (lwrite)  write(file_id,*) 'Nreact= ',reac,  'kf=', maxval(kf)
!

I stopped here, because because 
!

!print*,'p%TT=',p%TT
!AB: Here I find values of p%TT= 749.9975 2.5253 1.148736, etc.

Clearly, the chemistry is not valid at such low temperatures.
I guess this is just the pencil check which forces Natalia to implement
the physics of cold gases, and chemistry in the liquid phase.