Continued with Thomas the parallelization of the ADI scheme. Now the communications are embedeed in two separate subroutines in mpicomm.f90.
Also renamed the toto() subroutine in a more comprehensive way, that is,
ADI_Kconst_MPI(). 

BTW, the ADI is still located in the temperature_idealgas module because
we are continuing its parallelization. Once done, we plan to move it in
a separate module such that people that want to code their own implicit
time advance could start from it.

Boris.