Hi Fred, I guess you noticed that you checked into the google repository the full source code for each sample plus links such as
  getconf.csh -> /home/fred/pencil-code//bin/getconf.csh

The may not even work for you on all machines. It would also
be good to include reference data as in other samples, and to
have a format of the README file that is similar to those in
other samples, including email address, because they are being
extracted automatically if something goes wrong.
Axel