Coded a general initialize_implicit_physics() subroutine in the implicit_physics module. The idea is, first, to replace the too-restrictive former init_param_ADI() subroutine and, second, to adopt the general form used in any physics module with the succession
register_XXX

initialize_XXX

With this modification, a large part of the initialization of the ADI
parameters is at the correct place, that is, in the implicit_physics
module and not in the temperature_idealgas one. BTW, it will allow other
users to write their own implicit stuff in an easier way.

Boris.