- added a tridag_complex subroutine as we deal with a *complex* RHS in the anelastic case
- test the Poisson solver both in the fully periodic case (using the
classical inverse_laplacian subroutine coded in the Poisson module)
and in the semi-spectral case (i.e. the z direction is non-periodic).

The RHS is in both cases:

rhs(x,z) = -2*sin(x)*sin(z)

whom the solution is psi(x,y)=sin(x)*sin(z) [but also those with k=n/2
in the non-periodic case as k is then quantized due to the vertical
Dirichlet condition psi=0].

BUT I still have a f..... discrepancy near the surface between the
computed psi and the theoretical one in the semispectral case.