I'm experimenting with making the Makefile structure more clear.
I've removed some ancient targets, the dot stuff, an old 'ifc' debugging target
for radiation.

Furthermore I've made a separate list of files, called preprocess, that needs
to be preprocessed through mkcparam, since in the old system this is assumed to
be the same as the physics modules.

Please have a look (especially Wolfgang and Axel).

Also please try to copy the files the main src and auto-test. I don't want to
change the actual files before I am certain that it does not break anything.