Hi Wlad, I guess you have exported the initial condition of
  samples/2d-tests/globaldisc

into a separate initial condition sample, but it is not
currently checked in. I have commented out the 4 input variables
that are not currently available any more. I guess the problem is
similar with

  samples/dust_turb_globaldisk

and given the similarity between the two one wonders whether the
former is now obsolete. Also, I noticed that the pencil_consistency_check
finds a possible overcalculation of pencils rho, grho, and ugrho, as
well as pomx and pomy.
Axel