Sven, set mpi = `fgrep -c 'initialize_mpicomm: MPICOMM neighbors' src/start.x` doesn't work on the PDC cluster, so I reverted to previous line:
 set mpi = `fgrep -c 'mpicomm_init: MPICOMM neighbors' src/start.x`

Such changes in these scipts are quite dangerous, especially if they
break stuff on some machines. In this case, it concerned a machine
where the systems people change stuff at unexpected moments pretending
that it was a normal change. The last time I ran there was some weeks
ago, and there were some other changes with that machine. So now I
know that it would be premature to conclude that the systems people
did it again.