Corrected grammar and punctuation in comments.
I know that we used to to accept comments in lower case with no punctuation,
but the only stable solution is to use proper English (like in an e-mail or a
journal article). Otherwise we will not be able to find a common way to write
comments.

Dhruba, please add your signature to all the subroutines, including the day
they were programmed. One of them now says 'tony', which is probably not true.