I simplified the transp_xz subroutine written by Anders in the fall of 2006 by using a single call to MPI_SENDRECV_REPLACE in place of the former 4 calls MPI_SEND/MPI_RECV. I did that because the implicit runs done in parallel (that are using a lot transp_xz, in fact 3 successive calls at each timestep) didn't work otherwise above 8 cores due to an MPI blocking at large resolutions.
After many many attempts, I decided to simplify the MPI calls by
replacing:

proc x:
MPI_SEND
MPI_RECV

proc y:
MPI_RECV
MPI_SEND

by

proc x or y : 
MPI_SENDRECV_REPLACE

Only one buffer is then used and the four MPI calls SEND/RECV that were
done in an 'if/endif' structure are now embedded in an unique call to
MPI_SENDRECV_REPLACE. It results in an simplified version for transp_xz
that I checked a lot today and implicit things behave now perfectly
until at least 32 cores at the resolution 2048x512.

Of course the whole 32 auto-tests are OK but... to my knowledge none is
using transp_xz! I guess that this subroutine is actually only used in
the parallel version of the implicit solver (Thomas and me), or in the
semi-spectral poisson solver that is essential for the inhomogeneous
anelastic case (Piyali and me) or for a vertically-stratified disk
(Chao-Chin and Anders?).

Therefore I am commiting this change by hoping that everything will be
fine for the others...

Boris.