Axel, fixed your problem with power_spectra
-  if (lwrite_krms) then
+!  if (lwrite_krms) then
     call mpireduce_sum(k2m,k2m_sum,nk)
     call mpireduce_sum(nks,nks_sum,nk)
-  endif
+!  endif

For root-proc this variable was set false after writing but the other procs still had lwrite_krms=true

Another idea would be to distribute lwrite_krms every call of that subroutine.

But this depends on  the performance. Is it faster to call mpireduce even it is not needed or to send lwrite_krms to each proc.

Sven