Archaeology with the sample A3+chi11+Ra1e5:
- r8029 / 28aug07 
lbidiagonal_derij must be set by hand to F in the run.in file because
lbidiagonal_derij=T by default in cdata.f90 (and the reference.out file
was computed before this lbidiagonal_derij implementation)

- r10313 / 29jan09
the reference.out file must be updated after this revision committed by
Anders as the random numbers become pencilised with the initial conditions

Therefore Wolfgang: 

- I use now the default lbidiagonal_derij=T introduced by r8029 for this
  sample (to be coherent with other samples)

- I upgrade accordingly the reference.out that includes both: 
  * the lbidiagonal_derij=T flag (r8029)
  * the pencilisation of random numbers (r10313)

Are you OK with that?

Boris.