Long lists of modules to Use are written in alphabetical order; MR's latest update did not follow this convention.
Removed this commented out part:
-!  Set default to shearing sheet if lshear=.true. (even when Sshear==0).
-!
-!!!!      if (lshear) bcx(:)='she'
-!

I suppose that setting shear boundary conditions twice was really never
intended? So it was a genuine bug fix and the wrong lines should simply be
removed. In general I find it a bad habit to comment things out, since we have
SVN history to see past versions.