Added three calls to parallel_file_exists() for 1d-averages otherwise 2d runs compiled with gfortran may crash (at least with gfortran v4.1.2).
Further details: normally nnamex=nnamey=nnamez=0 in cdata.f90 but
we are using now Philippe's method with parallel_count_lines()
in register.f90. The pb is that if the usual 1d-average files
xyaver.in, xzaver.in & yzaver.in files are not present, then it
returned nnamex=nnamey=nnamez=-1. You see the pb: it crashes with a
gfortran-compiled code in xyaverages_z(), xzaverages_y() & yzaverages_x()
(diagnostics.f90) because one tries to initialise arrays with a negative
amount of memory as e.g.

real, dimension (nz,nprocz,nnamez) :: fsumz

Boris.