Galilean-invariant algorithm (that disk people know as "fargo") based on Fourier transforms. Works for 3D with x parallelization. Watch out for flows with sharp features, though, since Fourier transforms in this case lead to some ringing, that I tentatively identify as Gibbs phenomena. It's not aliasing (I tried de-aliasing but it was no remedy).
Real space shift works for non-parallel. Yet, because of 
Pencil's 3rd order timestep, the Fargo shift split into 
integer + fractional shift leads to an effective Courant 
number of 1.5. This is because the fractional shift 
(of max 0.5 cell) has to be done 3 times per full 
timestep. I had to split each subtimestep into 3 further 
sub-subtimesteps to do the fractional shift with proper 
accuracy. Still, it does not seem perfect yet - real space 
fargo needs more hyperviscosity than non-fargo advection. 

Moreover, as the derivative of a subtimestep is needed for the 
time-evolution of the next, in the fractional shift not only 
f but also df has to be discretized and advected. That 
introduces an extra discretization error that I'm not 
comfortable with. 

Comments welcome.