Influence of mesh hyperdiffusion in the timestep.
Strangely enough, they enter as an advective contribution, 
not as a diffusive one. 

According to Axel, that's because the term is linear in 
velocity, like an advection term. 

Clarifying (Axel, correct me if I am wrong in the lines below), 
the term is written as 

  d/dt rho  =  ksi / pi**5 d6rho/dx               (I)

Which can be seen to be equivalent to 

  d/dt rho = urms/Re * (dx/pi)**5 * d6rho/dx6     (II)

where 

  Re = urms * dx**5/(D3*pi**5)                    (III)

is the hyper mesh Reynolds number. 

So, ksi=urms/Re, and (I) is basically a modified 
advection. 

~x~

Also, modified slightly the influence on the timestep. Before, 
it had only the density contribution, which was being added to 
advec_cs2. Now they are separately added in quadrature and  
the square root added to maxadvec.