Renamed 'pc_load_file' to 'pc_load_dataset' in the hope that this name is more intuitive.
Added a procedure wrapper, so that this code can also be called from anywhere as procedure with arguments, too.

To use this code in scripts, one should only include the .inc file: '@pc_load_dataset.inc' or use it in the new procedure form. In interactive mode, one can use the procedure, while a simple '.r pc_load_dataset' is still possible, too.