warning: the people (bing means bingert?) that coded the superposed polytropic layers considered
 z<z1   : mpoly0
z1<z<z2 : mpoly1
z2<z    : mpoly2

while in entropy.f90 the historical choice is

 z<z1   : mpoly1
z1<z<z2 : mpoly0
z2<z    : mpoly2

as a consequence, the vertical profile for hcond is not the same than in
entropy.f90 (i.e. a diffrente combination of step calls).

In that respect, it may be interesting to adopt the same choice in entropy and
temperature_idealgas...