Natalie just noticed that the baroclinic sample now is crashing (I should have added it to the auto-test). I checked that the crash is due to a spurious shock developing in the boundary between processors. Strangely enough, it only occurs after 400 or so timesteps. It coincides with the instant when the y-boundaries are in opposite sides of the grid:
   -----------------
   |              .|
   |               |
   |X              |
   |               |
   |              .|
   -----------------

Tracing the problem back to the last working version, I found that it
was introduced in revision 16738, pertaining to what Anders described 
as a bug in the shock in shear periodic boundaries.

It turns out that the offending line is

   f(l1:l2,m1:m2,n1:n2,ishock) = tmp(l1:l2,m1:m2,n1:n2)*dxmax**2

whereas the old one

    f(:,:,:,ishock) = tmp*dxmax**2

works fine. It also works if one sets

   f(l1:l2,:,n1:n2,ishock) = tmp(l1:l2,:,n1:n2)*dxmax**2,

therefore, the problem is just with the y-boundary.

Considering that this is happening on a Friday afternoon, I just added
an ugly patch

       if (.not.lrewrite_shock_boundary) then
         f(l1:l2,m1:m2,n1:n2,ishock) = tmp(l1:l2,m1:m2,n1:n2)*dxmax**2
       else
         f(:,:,:,ishock) = tmp*dxmax**2
       endif

until I (or someone else) understands what is going on.