Hi Wlad and Anders and those who can help regarding the initial_condition module. I'm checking in some preparations for Andreas to do data assimilation by reading in 2 files (var.dat and another file of arbitrary name) into start, which case then produce a new initial condition. This is handled by setting, for example
  INITIAL_CONDITION = initial_condition/assimilate

The problem is that we want to read in two f arrays. In fact, start.f90
does already allocate an df array, and I am now simply using it, but
ideally I would like to call initial_condition_all(f,df) with 2 arguments.
I have commented this out right now, so that the auto-test works.
I am stuck now and would think that the simplest would be not to do it
through the initial_condition infra structure and to add some extra
assimilate infrastructure. What do you think?
Axel