Replaced notanumber implementation with one due to Sven and Philippe.
The latter is not extensively tested for different compilers and
optimization settings (at least not by myself; Sven and Phlippe did
test for a number of platforms). But with the previous implementation
I am getting false positives for subnormal numbers (e.g. 1e-40).
  This happens with GNU Fortran (Ubuntu/Linaro 4.5.2-8ubuntu4) 4.5.2 on
Ubuntu Natty, but Sven doesn't get the problem with the same compiler, on
the same OS. Weird...