BE CAREFUL, IMPORTANT COMMIT!!!
the default value cdt is now cdt=0.9 instead of cdt=0.4

see e.g. p528 in Canuto's book "Spectral methods: fundamentals in single
domains":

"The stability limits (on the imaginary axis) for these schemes are 1.73
for the third-order scheme and 3.34 for the fourth-order scheme..."

Please note that your sample must fail if the timestep is limited by
acoustic and you do not precise a particular cdt value in your run.in
file. By default the old cdt was cdt=0.4 therefore running your sample
with the new cdt=0.9 will lead to a different timestep, hence a
different reference.out file