Removed the 1/60. factor in the polar hyperdiffusion, used in cylindrical and spherical coordinates, that was introduced in revision r17303 (Sven, please check).
This factor was added to the RHS of the equations, but *not* 
to the respective contributions to max_diffus3, so that the 
timestep calculation was wrong. This was, for the same 
value of the hyperdiffusion, either the timestep was right but 
the Nyquist noise was not damped, or the Nyquist noise was damped 
but the timestep was too low. 

The solution is either to remove the 1./60 from the RHS, or to add 
that factor to the contributions to max_diffus3. I am for the former, 
since then it becomes more transparent that D' = D3 * k^4,
instead of D' = D3*k^4/60 - as well as saving a couple divisions.

This also shows that we need a sample to test the polar hyperdiffusion. 
I thought I had added that already, but apparently not. Doing so 
right away. 

The resolution-independent mesh-hyperdiffusion is not affected. But I 
think there is something pretty wrong with that formulation, and I don't 
like the way it affects the timestep via advection instead of via diffusion.