Removed 1/60 factor from the resolution-indepedent hyper-mesh viscosity, following discussion with Sven. A factor 1/60 was introduced in version 13333 in der6 with ignoredx=T, but the change was not propagated to the hyperdissipation coefficients that use that construction. That made polar hyperdiffusion and mesh-hyperdiffusion 1./60 weaker. Pumping up the coefficient by a factor 60 to yield the same dissipation was then unnecessarily decreasing the timestep.
Later on in revision 17303 that factor was removed from 
der6, but added to the hyperdissipation coefficients, thus 
not correcting the wrong behaviour. 

It should work now. 

ps: the factor 1/60 added to der6 with ignoredx was added for 
symmetry reasons, because all other derN have a non-unity numerical
factor with ignoredx. Question is, is the factor needed for the 
other derN? Maybe not - whoever coded it, please check.