Other stuff the pencil-check caught.
p%uxb is not needed for the advective gauge

But strangely enough (Anders, please check), it warns that p%bij "may not be needed". Yet, it is needed for calculating p%jj

          call curl_mn(p%bij,p%jj,p%bb)

To avoid the warning, this construction is added: 

          if (.not. lpencil(i_bij)) p%bij=0.0      ! Avoid warnings from pencil

Which looks a bit fishy. 

The same is going on in density for p%grho. It says it may not be needed, yet it is used to calculate p%ugrho in the continuity equation with ldensity_nolog=T.