Added a new routine 'output_form' to do formatted output with the distributed and collective IO-module. Removed some old code.
Bugfix for the collective module. It should now also work with very large data cubes - there was one problem that fseek can't jump to anything above a 32-bit integer and one other problem where the MPI library crashed with a Segmentation Fault when trying to send some 300 MB of data in one block.