Some of the code in pc_collect.f90 seems to not be needed for its functionality, in the sense that the code compiles and does its magic without them being invoked. These are
 register_modules
 input_globals
 initialize_modules
 choose_pencils

Actually, the code dies if register_modules is used in connection with global variables, such as the gravitational field in gravity_r.f90. To avoid this, this logical "lenforce_redundant" is added, that skips these pieces of problematic code. "lenforce_redundant" is set in run.in. Perhaps pc_collect should have its own namelist, such as "collect.in" or so.