Replaced lrandom_f_pencil_check by lpencil_check_no_zeros.
The code used to by default put random numbers in the f array for the pencil
check. However, I ran into a nasty error where the random temperature did not
affect the time-step, even though there was a pencil missing in the time-step
calculation.

In the new scheme, lpencil_check_no_zeros, any variable that is zero everywhere
is given random numbers instead, to avoid that composite pencils p%a=p%b*p%c
always pass the pencil check. Variables that are non-zero are kept as in the
initial condition.