Rewrote the boussinesq module as a density one (by starting from the nodensity.f90 one). I am clearly quite disappointed by the need to do that as we cannot actually use the boundcound module from the special one! I wrote first the boussinesq module as a special module but as I need to refresh the ghost zones for the velocity before taking its divergence, I must update before the BCs. Hence the BIG problem as a circular dependence appears: boundcond.f90 uses already the special module therefore we cannot call any subroutine enclosed in boundcond.f90 from the special module. Is it clear?
As a consequence, the boussinesq approximation lies on a given module
and I added a 

if (lanelastic) call boussinesq()

in timestep.f90, hydro.f90 and temperature_idealgas.f90.

For me this solution is without a doubt inelegant as we have in the
code the efficient special feature that can take care of that with its
special_after_timestep(), special_calc_hydro() or special_calc_entropy()
subroutines. But at the moment, because of the three special_boundconds()
calls in boundcond.f90, this is the only solution I found.