Another IO-module with collective IO based on hand-written torus-communication (first collect data along the y-direction, then along the x-direction).
Furthermore there are now variants of 'globalize_xy' and 'localize_xy' following the torus-communication concept, namely 'globalize_x_y' and 'localize_x_y'. These are supposed to be more efficient for large amounts of computing cores. Would be intresting to make a benchmark; I would estimate that this concept is already superior for nprocx=4 and nprocy>=4. If that is the case, the non-torus-type variants (including the 'collect_xy' IO module) can be deleted/replaced my the torus-variants.