if  (.not.ldensity.or.lanelastic) lpressuregradient_gas=.false. was overkill, because for .not.ldensity, we never need lpressuregradient_gas
Boussinesq approximation. Use Rayleigh number only with ltemperature.
Note: the buoyancy term is currently scaled with Ra*Pr, but Pr is also
regulated though mu and K, so Pr should eventually be eliminated.

Also, need to combine if (lboussinesq.and.ltemperature) then ...
and similarly if (ldensity.or.lanelastic) for mean momentum.