Removed the sharing of Prandtl number Pr between hydro and viscosity in the Boussinesq case. As Axel suggested, the user must rather set in the run.in file:
ivisc='simplified'

and gives the nu value. In the case of a timescale based on the radiative
diffusion, the viscous term is 

Pr * del2v

while the Prandtl number also enters in the momentum equation with a
buoyancy term as Ra * Pr * temp

where temp is the temperature perturbation. It means that the chosen
Prandtl number value set in the run.in file must also correspond to the
chosen nu value, at least when using this timescale.