"Render unto Caesar the things which are Caesar's...": Axel's remark about my (accidental) removal of former 2000 John Crowe contribution is absolutely justified, my apologies for John.
I first used his pendag() subroutine on laplacian systems and it did not
work, at least with my tests. As this subroutine was not used in the
code, this possible bug was transparent. I then recoded a new version
of this pendag() subroutine and tests were OK, the Boussinesq solver is
using it intensively when the fourth-order is required. I have had to
keep track of John Crowe's work in all cases, then I do.