Attempt on fixing the broken dust-vortex sample. The error message blames line 1361 of dustdensity, which is
  p%dndr(:,k)=-Imr*dndr_tmp(:,k)

Seemingly connected with the chemistry module. It looks like dndr_tmp is 
not initialized. It it set inside a subroutine that goes into fatal_error 
in the case in question. I initialized the array before the fatal error. 

This is a bit unexepcted, because the line in question in never called.