Hi Chao-Chin and Natalia,           I believe Chao-Chin has introduced multiple passive scalars in the code which also makes the pencil p%cc a two dimensional object with dimensions p%cc(1:nx,1:npscalar) . As the dimension of p%cc has changed, the dustdensity module has stopped compiling because it has been assuming that p%cc is a 1D pencil.  Normally a passive-scalar should not change any other module as it is be definition 'passive'. But it clearly makes some contribution in dustdensity. I have now commented this line.
Natalia, I  did not understand what you are trying to do with a passive scalar. Also you are multiplying it by an arbitrary number of 1e6 which is very bad practice. If you need such a number, it is better to give it a name and then have it as a defined variable. Could you two please communicate with each other to fix the problem ?