Added proof-of-concept tool 'pc_configtest' for checking the syntactic correctness of the "*.in" files.
This tool reads the "start.in" and "run.in" files and actually initializes all modules as it were a real run, but just with one processor (not using MPI). To build this tool, you can execute "pc_build tools" to build all tools or more specific "pc_build pc_configtest".