Use Test::NumericFileComparator for auto tests.
This eliminates code and improves diagnostics.

Note that the semantics of numerical comparisons is not identical to what
we had before: instead of doing tricky character manipulations on the
string representation of a number, we now simply use numerical comparison.
If no accuracy has been specified, 1.5 times the last digit in the largest
(by modulus) number in a column is used as tolerance.

The new semantics may break down if the required accuracy is comparable to
the accuracy of double-precision numbers (relative accuracy ~ 2.2e-16). In
that case, we should specify a slightly reduced precision.