Long-awaited F2003 functionality to read *.in-files by introduction of a new switchable 'file_io'-module.
* 'file_io_f2003' is still not finished - WILL BE RECOMMENDED, once it is finished
* 'file_io_f95' is "sloppy F95" with some ugly work-arounds to be HPC-friendly - RECOMMENDED FOR NOW (aka. former 'true_parallel_open')
* 'file_io_dist' is "strict F95" with distributed reading - NOT RECOMMENDED, because it is very HPC-unfriendly (aka. former 'fake_parallel_open')

The former switch 'FAKE_PARALLEL' has no effect anymore, you can delete it and simply switch modules within your "src/Makefile.local", if really really needed, but please report if one module doesn't work for you.

Thanks a lot for your cooperation and sorry for this monster-checkin, but it was not possible otherwise.