Implement switch loo_as_aux in module Hydro.
When loo_as_aux = .true., ioo as well as iox:ioz will be registered as
the index for a communicated auxiliary vector array in f-array, if not
already so.  The corresponding portion in f-array will be filled as the
vorticity field (curl of uu) each time before the mn-loop is entered.

Moreover, the registration of ioot and iuut is simplified.