Introduce parameter particle_radius.
This parameter can be used when module Particles_radius is not on, and
it means a constant radius for each constituent particle in particle
swarms.

Currently, it only has effects on calc_pencil_rep and particle removal
in solid cells.  In other places, fatal errors still occur when a
particle radius is required.