Modularize selection of particle-mesh method.
As a first step to centralize all particle-mesh related opeartions to
module Particles_map, move selecting the particle-mesh scheme to
Particles_map, either with or without particles blocks.  Note that the
particle-mesh schemes are independently implemented in the two cases.

This also implements the subroutine initialize_particles_map() so that
any particle-mesh related initialization can be done there.