Implemented particles_map_index.
This new implementation for the module Particles_map uses particle-
splitting approach to map the particles onto the grid.  In other words,
each cell contains the sub-particles contributed by nearby particles and
can be independently worked with.  Also, this module measure the
particle-mesh weighting against the index space instead of the real
space, such that non-equidistant grid is automatically scaled.