Add x-boundary conditions 'sf' and 'af'.
These are the equivalent of those implemented in z-direction.

'sf' is symmetric with respect to the domain boundary instead of the
first/last point.  Similarly, 'af' is antisymmetric with respect to the
domain boundary, which does not modify the value of the first/last point
as does boundary condition 'a'.